+++ src/utils/updater.py
@@ -15,7 +15,7 @@ try:
 except ImportError:
     ssl_context = ssl.create_default_context()
 
-VERSION = "1.5.26"
+VERSION = "1.5.27"
 
 def parse_version(v_str: str):
     """